+++ tests/test_hud_styling.py
@@ -62,7 +62,7 @@ class HudStylingTests(unittest.TestCase):
         def render(element=None, analysis=False):
             renderer = Renderer(*self.size, show_timecode=True, verbose=True, glow=0,
                                 analysis=analysis, analysis_target=analysis,
-                                geo_grid=True, target_outline=True, target_motif='triangles', target_label='TARGETING',
+                                geo_grid=True, target_outline=True, target_weak_spots=True, target_motif='triangles', target_label='TARGETING',
                                 subject_outline=True, subject_code=True, subject_labels=True,
                                 hud_blur_elements=None if element is None else element + '=8')
             image = Image.new('RGB', self.size)